add expression rule or add min max
