+++ fe/fe-core/src/main/java/org/apache/doris/nereids/rules/expression/ExpressionNormalizationAndOptimization.java
@@ -27,6 +27,7 @@ public class ExpressionNormalizationAndOptimization extends ExpressionRewrite {
                 ImmutableList.<ExpressionRewriteRule>builder()
                         .addAll(ExpressionNormalization.NORMALIZE_REWRITE_RULES)
                         .addAll(ExpressionOptimization.OPTIMIZE_REWRITE_RULES)
+                        .addAll(ExpressionOptimization.OPTIMIZE_REWRITE_RULES_OR_ADD_RANGE)
                         .build()
         ));
     }
