+++ fe/fe-core/src/test/java/org/apache/doris/nereids/rules/expression/SimplifyRangeTest.java
@@ -23,6 +23,7 @@ import org.apache.doris.nereids.analyzer.UnboundRelation;
 import org.apache.doris.nereids.analyzer.UnboundSlot;
 import org.apache.doris.nereids.parser.NereidsParser;
 import org.apache.doris.nereids.rules.analysis.ExpressionAnalyzer;
+import org.apache.doris.nereids.rules.expression.rules.OrAddMinMax;
 import org.apache.doris.nereids.rules.expression.rules.SimplifyRange;
 import org.apache.doris.nereids.trees.expressions.ExprId;
 import org.apache.doris.nereids.trees.expressions.Expression;
@@ -65,7 +66,7 @@ public class SimplifyRangeTest extends ExpressionRewrite {
     @Test
     public void testSimplify() {
         executor = new ExpressionRuleExecutor(ImmutableList.of(
-            bottomUp(SimplifyRange.INSTANCE)
+            bottomUp(SimplifyRange.INSTANCE, OrAddMinMax.INSTANCE)
         ));
         assertRewrite("TA", "TA");
         assertRewrite("TA > 3 or TA > null", "TA > 3 OR NULL");
@@ -204,7 +205,7 @@ public class SimplifyRangeTest extends ExpressionRewrite {
     @Test
     public void testSimplifyDate() {
         executor = new ExpressionRuleExecutor(ImmutableList.of(
-            bottomUp(SimplifyRange.INSTANCE)
+            bottomUp(SimplifyRange.INSTANCE, OrAddMinMax.INSTANCE)
         ));
         assertRewrite("AA", "AA");
         assertRewrite(
@@ -284,7 +285,7 @@ public class SimplifyRangeTest extends ExpressionRewrite {
     @Test
     public void testSimplifyDateTime() {
         executor = new ExpressionRuleExecutor(ImmutableList.of(
-            bottomUp(SimplifyRange.INSTANCE)
+            bottomUp(SimplifyRange.INSTANCE, OrAddMinMax.INSTANCE)
         ));
         assertRewrite("CA", "CA");
         assertRewrite(
@@ -369,7 +370,7 @@ public class SimplifyRangeTest extends ExpressionRewrite {
     @Test
     public void testAddMinMax() {
         executor = new ExpressionRuleExecutor(ImmutableList.of(
-            bottomUp(SimplifyRange.INSTANCE)
+            bottomUp(SimplifyRange.INSTANCE, OrAddMinMax.INSTANCE)
         ));
 
         assertRewrite("TA >= 10", "TA >= 10");
